Added hook for media search
